test(builder): separate builder tests into own files
