+++ test/builder/swagger.js
@@ -1,0 +1,27 @@
+// @flow
+import test from 'ava';
+
+import buildSwagger from '../../src/builder/swagger';
+
+test('handles additional properties', (t) => {
+  const TEST_DATA = {
+		Order: {
+      properties: { complete: { type: 'boolean', default: false } },
+      additionalProperties: { type: 'number' },
+			type: 'object',
+			xml: { name: 'Order' }
+		}
+	};
+  const TEST_RESPONSE = [
+    {
+      name: 'Order',
+      properties: {
+        complete: { exportTypes: undefined, importTypes: undefined, required: false, type: 'boolean' },
+        '[string]': { exportTypes: undefined, importTypes: undefined, required: false, type: 'number' }
+      }
+    }
+  ];
+
+  // $FlowFixMe
+  t.deepEqual(buildSwagger(TEST_DATA), TEST_RESPONSE);
+});
